# openapi.yaml
openapi: 3.0.3
info:
  title: Doctor.mx API
  description: |
    Doctor.mx - Plataforma integral de telemedicina para México.
    
    Esta API proporciona acceso a todas las funcionalidades de la plataforma Doctor.mx,
    incluyendo consultas médicas virtuales, gestión de citas, recetas digitales,
    análisis de IA, consentimientos digitales, y más.
    
    ## Autenticación
    La API utiliza Supabase Auth para la autenticación. Incluye el token de sesión
    en el header `Authorization: Bearer <token>`.
    
    ## Roles
    - **patient**: Pacientes que buscan atención médica
    - **doctor**: Médicos que brindan consultas
    - **admin**: Administradores del sistema
    
    ## Rate Limiting
    Las solicitudes están limitadas por IP y usuario autenticado.
    Headers de respuesta:
    - `X-RateLimit-Limit`: Límite de solicitudes
    - `X-RateLimit-Remaining`: Solicitudes restantes
    - `X-RateLimit-Reset`: Tiempo de reinicio
  version: 1.0.0
  contact:
    name: Doctor.mx Support
    email: soporte@doctor.mx
  license:
    name: Propietario
    url: https://doctor.mx/terms
servers:
  - url: https://doctor.mx
    description: Producción
  - url: https://staging.doctor.mx
    description: Staging
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/second-opinion:
    get:
      summary: Listar solicitudes
      description: Obtiene las solicitudes de segunda opinión del paciente
      tags:
        - Second Opinion
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecondOpinionRequest'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Crear solicitud
      description: Crea una nueva solicitud de segunda opinión
      tags:
        - Second Opinion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecondOpinionInput'
      responses:
        '201':
          description: Solicitud creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  price_cents:
                    type: integer
                  currency:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/referrals:
    get:
      summary: Listar referencias
      description: Obtiene referencias médicas del doctor
      tags:
        - Referrals
      parameters:
        - name: type
          in: query
          description: Tipo de referencia
          schema:
            type: string
            enum: [sent, received, all]
            default: all
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Lista de referencias
          content:
            application/json:
              schema:
                type: object
                properties:
                  referrals:
                    type: array
                    items:
                      $ref: '#/components/schemas/DoctorReferral'
                  stats:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Crear referencia
      description: Crea una nueva referencia médica
      tags:
        - Referrals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - patient_id
                - specialty_needed
                - reason
              properties:
                patient_id:
                  type: string
                  format: uuid
                specialty_needed:
                  type: string
                urgency:
                  type: string
                  enum: [emergency, urgent, routine]
                  default: routine
                reason:
                  type: string
                clinical_notes:
                  type: string
                receiving_doctor_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Referencia creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/upload/avatar:
    post:
      summary: Subir avatar
      description: Sube una foto de perfil
      tags:
        - Upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar subido
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de sesión de Supabase
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Cookie de sesión de Supabase
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-01-15T10:30:00Z"
...
    Currency:
      type: integer
      description: Monto en centavos
      example: 50000

    # ==================== PAGINATION ====================
...
    SecondOpinionType:
      type: string
      enum: [basic, specialist, panel]
    SecondOpinionStatus:
      type: string
      enum: [draft, pending_payment, paid, assigned, in_review, completed, cancelled]
    SecondOpinionInput:
      type: object
      required:
        - chief_complaint
      properties:
        type:
          $ref: '#/components/schemas/SecondOpinionType'
          default: basic
        chief_complaint:
          type: string
          minLength: 10
        current_diagnosis:
          type: string
        current_treatment:
          type: string
        medical_history:
          type: string
        allergies:
          type: string
        medications:
          type: string
        questions:
          type: array
          items:
            type: string
    SecondOpinionRequest:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        patient_id:
          $ref: '#/components/schemas/UUID'
        type:
          $ref: '#/components/schemas/SecondOpinionType'
        chief_complaint:
          type: string
        current_diagnosis:
          type: string
          nullable: true
        current_treatment:
          type: string
          nullable: true
        medical_history:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/SecondOpinionStatus'
        price_cents:
          $ref: '#/components/schemas/Currency'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'

    # ==================== REFERRALS ====================
    DoctorReferral:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        referring_doctor_id:
          $ref: '#/components/schemas/UUID'
        receiving_doctor_id:
          $ref: '#/components/schemas/UUID'
          nullable: true
        patient_id:
          $ref: '#/components/schemas/UUID'
        specialty_needed:
          type: string
        urgency:
          type: string
          enum: [emergency, urgent, routine]
        reason:
          type: string
        clinical_notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, completed, declined, expired]
        expires_at:
          $ref: '#/components/schemas/Timestamp'
        created_at:
          $ref: '#/components/schemas/Timestamp'

    # ==================== FOLLOWUPS ====================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Bad Request"
            message: "Datos de entrada inválidos"
            details:
              field: "email"
              message: "Email inválido"
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Se requiere autenticación"
    Forbidden:
      description: Acceso denegado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Forbidden"
            message: "No tiene permisos para esta acción"
...